Minor fixes nor fixes# Your branch is up to date with 'origin/master'.

+++ Framework/main/fading.ino
@@ -5,9 +5,10 @@ void startFading() {
 }
 
 void fade() {
-  double fadeVal = -1 * ((millis() - startFade) / getVar("fading_time").value - 1);
+  double fadeVal = -1 * ((millis() - startFade) / double(getVar("fading_time").value) - 1);
+
   if (fadeVal > 0) {
-    showColor(fadeVal * currColor.red, fadeVal* currColor.green, fadeVal * currColor.blue);
+    showColor(fadeVal * currColor.red, fadeVal * currColor.green, fadeVal * currColor.blue);
   } else {
     setState(OFF);
     if (lastOn == id) {
